requirements.txt
pandas
seaborn
numpy
matplotlib
seaborn
scikit-learn
flask
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gspread
psycopg2-binary




